# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Dondok API
  version: 0.10.0
  description: 돈독 프론트엔드와 백엔드 사이의 HTTP 계약
  license:
    name: All rights reserved
    identifier: LicenseRef-Proprietary
servers:
  - url: /
security: []
paths:
  /api/auth/csrf:
    get:
      summary: CSRF token 조회
      operationId: getCsrfToken
      tags: [Authentication]
      responses:
        "200":
          description: 현재 브라우저 세션용 CSRF token
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CsrfToken" }
        "400": { $ref: "#/components/responses/BadRequest" }
  /api/auth/login-ids/{loginId}/availability:
    get:
      summary: 로그인 아이디 중복 확인
      operationId: getLoginIdAvailability
      tags: [Authentication]
      parameters:
        - name: loginId
          in: path
          required: true
          schema: { type: string, pattern: "^[A-Za-z0-9._-]{4,30}$" }
      responses:
        "200":
          description: 아이디 사용 가능 여부
          content:
            application/json:
              schema:
                type: object
                required: [available]
                properties:
                  available: { type: boolean }
        "400": { $ref: "#/components/responses/BadRequest" }
  /api/auth/sign-up:
    post:
      summary: 회원가입
      operationId: signUp
      tags: [Authentication]
      parameters: [{ $ref: "#/components/parameters/CsrfHeader" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SignUpRequest" }
      responses:
        "202":
          description: 가입 완료, 이메일 인증 대기
          content:
            application/json:
              schema:
                type: object
                required: [email]
                properties:
                  email: { type: string, format: email }
        "400": { $ref: "#/components/responses/BadRequest" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/auth/email-verifications:
    post:
      summary: 이메일 인증
      operationId: verifyEmail
      tags: [Authentication]
      parameters: [{ $ref: "#/components/parameters/CsrfHeader" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TokenRequest" }
      responses:
        "204": { description: 이메일 인증 완료 }
        "400": { $ref: "#/components/responses/BadRequest" }
  /api/auth/session:
    post:
      summary: 세션 로그인
      operationId: login
      tags: [Authentication]
      parameters: [{ $ref: "#/components/parameters/CsrfHeader" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [loginId, password]
              properties:
                loginId: { type: string }
                password: { type: string, format: password }
      responses:
        "200":
          description: 로그인 완료
          headers:
            Set-Cookie:
              schema: { type: string }
              description: HttpOnly DONDOK_SESSION cookie
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SessionUser" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    delete:
      summary: 현재 세션 로그아웃
      operationId: logout
      tags: [Authentication]
      security: [{ cookieAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/CsrfHeader" }]
      responses:
        "204": { description: 현재 세션 로그아웃 }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/auth/sessions:
    delete:
      summary: 모든 세션 로그아웃
      operationId: logoutAll
      tags: [Authentication]
      security: [{ cookieAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/CsrfHeader" }]
      responses:
        "204": { description: 모든 세션 로그아웃 }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/auth/me:
    get:
      summary: 현재 로그인 사용자 조회
      operationId: getCurrentUser
      tags: [Authentication]
      security: [{ cookieAuth: [] }]
      responses:
        "200":
          description: 현재 로그인 사용자
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SessionUser" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/auth/password-resets:
    post:
      summary: 비밀번호 재설정 메일 요청
      operationId: requestPasswordReset
      tags: [Authentication]
      parameters: [{ $ref: "#/components/parameters/CsrfHeader" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email, maxLength: 320 }
      responses:
        "202": { description: 계정 존재 여부와 무관하게 동일한 응답 }
        "400": { $ref: "#/components/responses/BadRequest" }
  /api/auth/password-resets/confirm:
    post:
      summary: 비밀번호 재설정 확정
      operationId: confirmPasswordReset
      tags: [Authentication]
      parameters: [{ $ref: "#/components/parameters/CsrfHeader" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, newPassword]
              properties:
                token: { type: string, maxLength: 200 }
                newPassword: { type: string, format: password, minLength: 10, maxLength: 128 }
      responses:
        "204": { description: 비밀번호 변경과 기존 전체 세션 만료 완료 }
        "400": { $ref: "#/components/responses/BadRequest" }
...
  /api/categories:
    get:
      summary: 현재 가계부의 활성 분류 목록
      operationId: getCategories
      tags: [Category]
      security: [{ cookieAuth: [] }]
      parameters:
        - name: kind
          in: query
          required: true
          schema: { $ref: "#/components/schemas/CategoryKind" }
      responses:
        "200":
          description: 정렬된 공동 분류
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Category" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    post:
      summary: 공동 분류 추가
      operationId: createCategory
      tags: [Category]
      security: [{ cookieAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/CsrfHeader" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateCategoryRequest" }
      responses:
        "201":
          description: 생성된 분류의 authoritative 상태
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Category" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/categories/{categoryId}:
    parameters:
      - name: categoryId
        in: path
        required: true
        schema: { type: string, format: uuid }
    put:
      summary: expected version 기준 공동 분류 이름 변경
      operationId: updateCategory
      tags: [Category]
      security: [{ cookieAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/CsrfHeader" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateCategoryRequest" }
      responses:
        "200":
          description: 변경된 분류의 authoritative 상태
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Category" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "412": { $ref: "#/components/responses/PreconditionFailed" }
    delete:
      summary: 연결 거래를 같은 방향 기타로 이동한 뒤 공동 분류 보관
      operationId: archiveCategory
      tags: [Category]
      security: [{ cookieAuth: [] }]
      parameters:
        - { $ref: "#/components/parameters/CsrfHeader" }
        - name: expectedVersion
          in: query
          required: true
          schema: { type: integer, format: int64, minimum: 0 }
      responses:
        "200":
          description: 실제 거래 이동 결과
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ArchiveCategoryResult" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "412": { $ref: "#/components/responses/PreconditionFailed" }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: DONDOK_SESSION
  parameters:
    CsrfHeader:
      name: X-CSRF-TOKEN
      in: header
      required: true
      schema: { type: string }
...
  schemas:
    CsrfToken:
      type: object
      required: [headerName, parameterName, token]
      properties:
        headerName: { type: string, const: X-CSRF-TOKEN }
        parameterName: { type: string }
        token: { type: string }
    SignUpRequest:
      type: object
      required: [loginId, displayName, email, password]
      properties:
        loginId: { type: string, pattern: "^[A-Za-z0-9._-]{4,30}$" }
        displayName: { type: string, minLength: 1, maxLength: 100 }
        email: { type: string, format: email, maxLength: 320 }
        password: { type: string, format: password, minLength: 10, maxLength: 128 }
    TokenRequest:
      type: object
      required: [token]
      properties:
        token: { type: string, maxLength: 200 }
    SessionUser:
      type: object
      required: [userId, loginId, displayName, email]
      properties:
        userId: { type: string, format: uuid }
        loginId: { type: string }
        displayName: { type: string }
        email: { type: string, format: email }
...
    CategoryKind:
      type: string
      enum: [INCOME, EXPENSE]
    Category:
      type: object
      required: [categoryId, kind, systemCode, name, isFallback, transactionCount, version]
      properties:
        categoryId: { type: string, format: uuid }
        kind: { $ref: "#/components/schemas/CategoryKind" }
        systemCode:
          oneOf:
            - { type: string }
            - { type: "null" }
        name: { type: string }
        isFallback: { type: boolean }
        transactionCount: { type: integer, format: int64, minimum: 0 }
        version: { type: integer, format: int64, minimum: 0 }
    CreateCategoryRequest:
      type: object
      required: [kind, name]
      properties:
        kind: { $ref: "#/components/schemas/CategoryKind" }
        name: { type: string, minLength: 1, maxLength: 100 }
    UpdateCategoryRequest:
      type: object
      required: [name, expectedVersion]
      properties:
        name: { type: string, minLength: 1, maxLength: 100 }
        expectedVersion: { type: integer, format: int64, minimum: 0 }
    ArchiveCategoryResult:
      type: object
      required:
        - categoryId
        - fallbackCategoryId
        - fallbackCategoryName
        - remappedTransactionCount
        - firstOccurredOn
        - lastOccurredOn
      properties:
        categoryId: { type: string, format: uuid }
        fallbackCategoryId: { type: string, format: uuid }
        fallbackCategoryName: { type: string }
        remappedTransactionCount: { type: integer, format: int64, minimum: 0 }
        firstOccurredOn:
          oneOf:
            - { type: string, format: date }
            - { type: "null" }
        lastOccurredOn:
          oneOf:
            - { type: string, format: date }
            - { type: "null" }
...
    Problem:
      type: object
      required: [status, title, detail, errorCode, correlationId, timestamp, fieldErrors]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string, format: uri-reference }
        errorCode: { type: string }
        correlationId: { type: string }
        timestamp: { type: string, format: date-time }
        fieldErrors:
          type: array
          items:
            type: object
            additionalProperties: true
  responses:
    BadRequest:
      description: 입력값 또는 token 오류
      content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } }
    Unauthorized:
      description: 인증 실패
      content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } }
    Forbidden:
      description: 이메일 미인증 또는 CSRF 오류
      content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } }
    Conflict:
      description: 현재 상태와 충돌하는 요청
      content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } }
    NotFound:
      description: 현재 사용자가 접근할 수 있는 리소스가 없음
      content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } }
    PreconditionFailed:
      description: 편집 시작 후 다른 세션이 리소스를 수정함
      content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } }